Prison compatibility: Added support for block metadata customModelData. This is only compatible with spigot 1.14.x and higher.
