+++ prison-spigot/src/main/java/tech/mcprison/prison/spigot/compat/Spigot_1_13.java
@@ -1,21 +1,3 @@
-/*
- *  Prison is a Minecraft plugin for the prison game mode.
- *  Copyright (C) 2017 The Prison Team
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
 package tech.mcprison.prison.spigot.compat;
 
 import org.bukkit.Effect;
@@ -30,13 +12,8 @@ import tech.mcprison.prison.spigot.SpigotUtil;
 import tech.mcprison.prison.spigot.block.SpigotItemStack;
 import tech.mcprison.prison.spigot.inventory.SpigotPlayerInventory;
 
-/**
- * Spigot 1.9, 1.10, and 1.11.
- *
- * @author Faizaan A. Datoo
- */
-public class Spigot19 
-	extends Spigot19GUI
+public class Spigot_1_13
+	extends Spigot_1_13_GUI
 	implements Compatibility {
 
     @Override 
@@ -62,7 +39,7 @@ public class Spigot19
     public ItemStack getItemInMainHand(PlayerInventory playerInventory) {
     	return playerInventory.getItemInMainHand();
     }
-    
+   
 	@Override
     public SpigotItemStack getPrisonItemInMainHand(PlayerInteractEvent e) {
     	return SpigotUtil.bukkitItemStackToPrison( getItemInMainHand( e ) );
@@ -101,18 +78,19 @@ public class Spigot19
     }
 
     @Override
-    public void setItemInMainHand(Player p, ItemStack itemStack) {
+    public void setItemInMainHand(Player p, ItemStack itemStack){
         p.getInventory().setItemInMainHand(itemStack);
     }
-
+    
     @Override
     public void setItemStackInOffHand( SpigotPlayerInventory inventory, SpigotItemStack itemStack ) {
-
+    	
     	ItemStack iStack = itemStack == null ? null : itemStack.getBukkitStack();
     	
     	((org.bukkit.inventory.PlayerInventory) inventory.getWrapper())
-    	.setItemInOffHand( iStack );
+    			.setItemInOffHand( iStack );
     }
+    
     @Override 
     public void playIronDoorSound(Location loc) {
         loc.getWorld().playEffect(loc, Effect.IRON_DOOR_TOGGLE, null);
@@ -149,5 +127,4 @@ public class Spigot19
 		
 		player.playSound(player.getLocation(), getEntityItemBreakSound(), 1.0F, 0.85F);
 	}
-
 }
