+++ prison-spigot/src/main/java/tech/mcprison/prison/spigot/compat/Spigot_1_13_GUI.java
@@ -2,8 +2,8 @@ package tech.mcprison.prison.spigot.compat;
 
 import org.bukkit.event.inventory.InventoryEvent;
 
-public abstract class Spigot113GUI 
-		extends Spigot113Blocks 
+public abstract class Spigot_1_13_GUI 
+		extends Spigot_1_13_Blocks 
 		implements CompatibilityGUI {
 
     @Override
