+++ prison-spigot/src/main/java/tech/mcprison/prison/spigot/compat/Spigot_1_14.java
@@ -1,0 +1,6 @@
+package tech.mcprison.prison.spigot.compat;
+
+public class Spigot_1_14
+	extends Spigot_1_14_Blocks {
+
+}
