+++ prison-spigot/src/main/java/tech/mcprison/prison/spigot/compat/Spigot_1_9_GUI.java
@@ -2,8 +2,8 @@ package tech.mcprison.prison.spigot.compat;
 
 import org.bukkit.event.inventory.InventoryEvent;
 
-public class Spigot19GUI 
-		extends Spigot19Player 
+public class Spigot_1_9_GUI 
+		extends Spigot_1_9_Player 
 		implements CompatibilityGUI {
 	
     @Override
